scripts,advisor: Tweak launcher script
